Fix behavior of Apply for new Pipeline Rules
## Motivation
Prior to this change, if a user was clicking 'Apply' when creating a
rule a new rule was created but the page was still on the new page which
would lead into an error when clicking 'Save & Close' next.

## Description
This change will change the browsers url to the edit page after clicking
apply which will fix the state after coming from a new rule page.